wip readding icons since they didnt save

+++ code/pawn/Gun.cs
@@ -10,7 +10,7 @@ public class Gun : ModelEntity {
         SetModel(model);
         muzzle = Model.GetAttachment("muzzle")?.Position ?? Vector3.Zero;
     }
-    public void Fire(Pawn owner, TraceResult tr, int damage, Action react) {
+    public void Fire(Pawn owner, TraceResult tr, float damage, Action react) {
         DebugOverlay.Line(Position + muzzle * owner.Scale * owner.Rotation, tr.EndPosition, 0.1f, true);
 
         PlaySound("sounds/fire.sound");
